% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisComplex.R
\name{complexFunctionPlot}
\alias{complexFunctionPlot}
\title{complexFunctionPlot}
\usage{
complexFunctionPlot(...)
}
\arguments{
\item{...}{All parameters \code{\link{phasePortrait}} can be called with,
using exactly the same definitions.}
}
\description{
This function is just a wrapper for \code{phasePortrait} in order to
guarantee compatibility with earlier versions. Don't use
\code{complexFunctionPlot} when you are writing new scripts. It is
called with exactly the same parameters as \code{\link{phasePortrait}}
}
