% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{juliaNormal}
\alias{juliaNormal}
\title{Julia iteration with a given number of steps}
\usage{
juliaNormal(z, c, R_esc, itDepth = 500L)
}
\description{
Julia iteration with a given number of steps
}
\seealso{
Other fractals: 
\code{\link{mandelbrot}()}
}
\concept{fractals}
