% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phasePortraitBw.R
\name{phasePortraitBw}
\alias{phasePortraitBw}
\title{Black-and-white phase portraits}
\usage{
phasePortraitBw(
  FUN,
  moreArgs = NULL,
  xlim,
  ylim,
  invertFlip = FALSE,
  res = 150,
  blockSizePx = 2250000,
  tempDir = NULL,
  nCores = parallel::detectCores(),
  pType = "pma",
  pi2Div = 9,
  logBase = exp(2 * pi/pi2Div),
  argOffset = 0,
  bwCols = c("black", "grey95", "grey"),
  asp = 1,
  deleteTempFiles = TRUE,
  noScreenDevice = FALSE,
  autoDereg = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Black-and-white phase portraits
}
\examples{

}
